Create test_map.html
Implementing a map with a 25 second "cool down period" after the transition from location page back to the map page. The goal will be to lessen the possibility of inadvertent bouncing between the two pages.

This should allow tour goers to inadvertently travel through the geofence within the cool down (on their way to the next site) without having to worry about a page change.

Testing @ 25 seconds to see if that is enough time (or too much).