Update tt.html
Ok. This needs to be tested, but for all intents and purposes it is a finished product. The user moves toward a location marker --> within the distance threshold a location pop-up opens --? user see pic of historic location and listens to audio file discussing history --> user answers trivia question from audio file --> returned to the map page to begin toward the next tour marker. Locations load in order one at a time and a congratulations pop-up lets the user know the tour has ended and returns them to the main page (might change that to the tour page). There will be optimizations, but I think this is ready to go -- or at least close.