Update tt.html
Updated the checkAnswer function to give a different message to users when they answer the trivia question after the last tour location's audio file. The vanilla "Moving to the next location." didn't fit well here. Changing it to You've completed the tour. Will change the showCongratulations function message next to account for redundancies.